requirements.txt
python-slugify
kegscraper
playwright
